cli: Add `idl type` command
